Hopefully fixed GDI bug, Added default and dark skin. Changed main screen to use SortableListView.
Instead of passing around bitmaps, now mostly deals with byte arrays and
converts only when necessary, caching the image where possible.

Two skins are available. First time I've ever tried anything like
skinning so it's probably hacky, and I'm not sure how much I like it.
Option is there is anyone wants to use it, it is currently *not*
persistent.

Main screen is now the same type of sortable list view as others.