Amended applying skin logic
Implemented an 'ISkinnableControl' containing an 'ApplySkin' method.
Moved all the logic for this into the controls themselves rather than
all the logic being in one method on the form.

Amended a few things just to fit in with the coding style in place, will
review these again later on.

Added credits to the 'About' menu and readme.