Add german localization
This adds localization for the german language. It uses the formal form (Sie).
I translated everything in the Strings-file, but it seems not every control in the UI uses localized strings. I will revisit them, if something changes in this regard.
One minor adjustment had to be made to:
This application will ALWAYS be free, open source, and supported to the best of my ability.
Which had to be translated so is means something like:
This application will ALWAYS be free and open source and I will continue to support it to my best intend.